@@ -1,11 +1,11 @@
 name: flutter_app_lock
 description: A Flutter package for showing a lock screen on app open and app pause.
-version: 4.0.0
+version: 4.0.1
 homepage: https://github.com/tomalabaster/flutter_app_lock
 
 environment:
   sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=3.0.0"
+  flutter: ">=3.16.0"
 
 dependencies:
   flutter:
@@ -14,4 +14,4 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
